Moved around scripts, deleted unused scripts
