+++ Assets/Scripts/Dashboard/Tutorial/LobbyDefault.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
+using Anivision.PlayerInteraction;
 using TMPro;
 using UnityEngine;
 using Anivision.SceneManagement;
