--- Assets/Scripts/SceneReference.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 215a439e197f7bd4d91dbff809331ffd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
