--- Assets/Scripts/leftControls.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a7a11d0bd85b34549ad9a0cd8823d2e4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
