--- Assets/Scripts/returnToLobby.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class returnToLobby : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        SceneManager.LoadScene("TutorialScene");
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
