--- Assets/Scripts/returnToLobby.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 495935e0a06a1cd4f95cc5a0a99b7a05
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
