--- Assets/Scripts/startButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 82ef9e8f23d887044a3016ff63db9134
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
