--- Assets/Scripts/tutorialOld.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fce814e85f4dd054cadd3adde11b3cf1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
