--- Assets/Scripts/tutorialOld/ButtonPressTutorial.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ae7e8edbdfd26304b83ab9dc10630561
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
