--- Assets/Scripts/tutorialOld/GrabTutorial.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 21cd0a3045cbe3444b58c9894ed7633e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
