--- Assets/Scripts/tutorialOld/SwitchVisionsTutorial.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e2b0f073cc01bb444ad3878f9ebe34b9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
