--- Assets/Scripts/tutorialOld/TeleportTutorial.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2796155acc39a224ca67119aaca93a24
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
