--- Assets/Scripts/tutorialOld/TutorialBaseClass.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e106893da06c260478528431ad3a99b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
