--- Assets/Scripts/tutorialOld/TutorialGameController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7bf9ee2e576d5954a8de7e1c3784ffbb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
